Share function
move dublicate function and share with others
