+++ script/compare.js
@@ -1,62 +1,4 @@
-function getSVGIcon(iconCode) {
-    let iconPath;
-    switch (iconCode) {
-        case '01d':
-            iconPath = 'icons/animated/day.svg';
-            break;
-        case '01n':
-            iconPath = 'icons/animated/night.svg';
-            break;
-        case '02d':
-            iconPath = 'icons/animated/cloudy-day-1.svg';
-            break;
-        case '02n':
-            iconPath = 'icons/animated/cloudy-night-1.svg';
-            break;
-        case '03d':
-            iconPath = 'icons/animated/cloudy-day-2.svg';
-            break;
-        case '03n':
-            iconPath = 'icons/animated/cloudy-night-2.svg';
-            break;
-        case '04d':
-            iconPath = 'icons/animated/cloudy-day-3.svg';
-            break;
-        case '04n':
-            iconPath = 'icons/animated/cloudy-night-3.svg';
-            break;
-        case '09d':
-            iconPath = 'icons/animated/rainy-1.svg';
-            break;
-        case '09n':
-            iconPath = 'icons/animated/rainy-4.svg';
-            break;
-        case '10d':
-            iconPath = 'icons/animated/rainy-2.svg';
-            break;
-        case '10n':
-            iconPath = 'icons/animated/rainy-5.svg';
-            break;
-        case '11d':
-        case '11n':
-            iconPath = 'icons/animated/thunder.svg';
-            break;
-        case '13d':
-            iconPath = 'icons/animated/snowy-1.svg';
-            break;
-        case '13n':
-            iconPath = 'icons/animated/snowy-5.svg';
-            break;
-        case '50d':
-        case '50n':
-            iconPath = 'icons/animated/cloudy.svg';
-            break;
-        default:
-            iconPath = 'icons/animated/cloudy.svg';
-            break;
-    }
-    return iconPath;
-}
+import getSVGIcon from './global.js'
 
 function toggleDropdown(dropdownContentSelector) {
     const dropdownContent = document.querySelector(dropdownContentSelector);
